Update styling of links
This commit brings the link style in line with the assignment
directions. Using flex on the links I arranged them horizontally, then
refering back to a previous assignment, arranged them with space between
and then removed the list styling as well as link styling. Now just need
to force header and footer to top and bottom of the page, leaving the
content box to expand and fill full page.